requirements.txt
greenlet==3.2.4
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pycairo==1.29.0
PyGObject==3.54.5
pytest==8.3.5
python-dotenv==1.1.0
requirements-parser==0.13.0
SQLAlchemy==2.0.40
typing_extensions==4.15.0
